wip: move trail-graphql functionaity into fastify plugin
