+++ packages/trail-fastify-graphql-plugin/lib/graphql.js
