--- packages/trail-graphql/config/default.js
@@ -1,13 +1,0 @@
-'use strict'
-
-module.exports = {
-  db: {
-    host: 'localhost',
-    port: 5432,
-    database: 'trails',
-    username: 'postgres',
-    password: 'postgres',
-    poolSize: 10,
-    idleTimeoutMillis: 30000
-  }
-}
