--- packages/trail-graphql/config/test.js
@@ -1,7 +1,0 @@
-'use strict'
-
-module.exports = {
-  db: {
-    database: 'trails_test'
-  }
-}
