--- packages/trail-graphql/config/travis.js
@@ -1,8 +1,0 @@
-'use strict'
-
-module.exports = {
-  db: {
-    database: 'trails_test',
-    password: ''
-  }
-}
