--- packages/trail-graphql/jest.config.js
@@ -1,3 +1,0 @@
-'use strict'
-
-module.exports = require('../../jest.config')
