--- packages/trail-graphql/package.json
@@ -1,50 +1,0 @@
-{
-  "name": "@nearform/trail-graphql",
-  "version": "3.0.0",
-  "description": "Graphql adapter for audit trail logging service",
-  "license": "MIT",
-  "author": "nearForm Ltd",
-  "contributors": [
-    "Julian Goacher (https://github.com/juliangoacher)"
-  ],
-  "homepage": "https://github.com/nearform/trail#readme",
-  "keywords": [
-    "audit",
-    "log",
-    "logging",
-    "trail",
-    "graphql"
-  ],
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/nearform/trail.git"
-  },
-  "bugs": {
-    "url": "https://github.com/nearform/trail/issues"
-  },
-  "engines": {
-    "node": ">=12.0.0"
-  },
-  "publishConfig": {
-    "access": "public"
-  },
-  "main": "lib/index.js",
-  "scripts": {
-    "coverage": "lab -c test",
-    "depcheck": "npx depcheck --ignores=config",
-    "test": "lab test"
-  },
-  "dependencies": {
-    "@nearform/trail-core": "^3.0.0",
-    "graphql": "^14.6.0",
-    "graphql-jit": "0.4.2",
-    "graphql-tools": "4.0.7",
-    "graphql-type-json": "^0.3.1",
-    "luxon": "1.22.0"
-  },
-  "devDependencies": {
-    "@hapi/code": "^8.0.1",
-    "@hapi/lab": "^22.0.3",
-    "sinon": "^8.1.1"
-  }
-}
